feat(app-workspace): GUI workspace with goal flow + wiki preview
Replace WorkspaceStub with a real Workspace shell after the user opens
a vault from the Lobby. Three tabs (Goals / Wiki / Quiz), Goal lifecycle
spawn / cancel / stream from the GUI, and a read-only wiki preview with
client-side wikilink resolution. Closes v3-app-workspace-goal change.

Highlights:

* New IPC commands (9 -> 15): spawn_goal / cancel_goal / list_runs /
  get_run_detail / list_wiki_pages / read_wiki_page. spawn_goal runs in
  a std::thread::Builder + catch_unwind so verb panics never poison the
  active_runs map. A new goal-terminal Tauri event channel notifies
  frontend when the background thread exits so the UI can flip activeRun
  back to null + re-list runs + invalidate wiki cache without polling.
* RunId = the run's started_at slug (per design); one active goal run
  per vault enforced at both layers (frontend Run button disabled while
  activeRun != null; backend rejects with AppError::Invalid).
* Interrupted run detection: events-*.jsonl with no matching RunLog row
  synthesizes a virtual outcome="interrupted" entry at list_runs time.
* Zustand stores: useGoalsStore (runs / activeRun / spawn / cancel /
  refresh) + useWikiStore (pages / currentPath / body / loadPage /
  listPages). Both subscribe to goal-terminal for cross-store refresh.
* Stream rendering aligned with CLI render/stream_event.rs: emoji
  leaders (Tools render as the Write/Edit wrench/pencil, Thought renders
  with the thinking emoji, banners use the bus / target / sync / commit
  / done emojis). Thought events fold inline at their original timeline
  position (not buffered to the bottom) with multi-line collapse.
* Done detail: per-phase Activity summary + Covered pages grouped by
  SpawnStart/SpawnEnd verb boundaries (goal phase vs fix phase), a
  collapsible Run details full event replay, and clickable Covered
  pages rendered with frontmatter title (slug fallback).
* Wiki preview: react-markdown + remark-gfm with Tailwind-styled
  components (max-width 720px, system font stack, sunken code blocks,
  bordered tables, Obsidian-style purple links). Wikilinks display the
  page title; unresolvable wikilinks render as dimmed spans with
  "Page not found" tooltip. tauri-plugin-opener adds a click-to-reveal
  on the sidebar vault path.
* CLI alignment fixes: VerbBanner SyncDone / LintDone elapsed_ms u128
  -> u64 (default serde_json deserialize support; no wire impact),
  verb::goal::run_goal now emits SpawnStart / SpawnEnd { verb: Fix } so
  GUI phase grouping can attribute fix-loop tool calls correctly.
* WindowControls overlap fixes: all Workspace top bars reserve right
  padding for the frameless controls and mark themselves as drag
  regions; LoadingOverlay is also drag-able so the user can move the
  window while "boarding the bus".
* Backlog notes captured for two follow-up changes (skill bundles
  vault-only, git-context-tool with PII coverage) under docs/.

37/37 tasks complete. Tests: 71 + 369 + 110 Rust pass, 201 Vitest pass,
typecheck clean, spectra validate + analyze clean (0 critical / 0 warn,
3 pre-existing low-severity language suggestions).

+++ codebus-cli/src/commands/init.rs
@@ -122,7 +122,7 @@ fn handle_event(event: &InitEvent<'_>, debug: bool, render_opts: &RenderOptions)
                 Banner::SyncDone {
                     files: summary.files,
                     mib: bytes_to_mib(summary.bytes),
-                    elapsed_ms: *elapsed_ms,
+                    elapsed_ms: *elapsed_ms as u64,
                 },
                 render_opts,
             );
